divided report dialog and report icon
